pyproject.toml
# https://python-poetry.org/docs/pyproject/

# PROJECT

[tool.poetry]
name = "fastapi_autogen_team"
version = "0.1.0"
description = "Autogen team with fastapi with opanapi compatible"
authors = ["lgcorzo <lgcorzo@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["mlops", "python", "package", "fastapi"]
packages = [{ include = "fastapi_autogen_team", from = "src" }]

# DEPENDENCIES

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.115.11"
ag2 = {extras = ["openai"], version = "^0.8.3"}
flaml = "^2.3.4"
opentelemetry-api = "^1.31.1"
opentelemetry-sdk = "^1.31.1"
opentelemetry-instrumentation-fastapi = "^0.52b1"
opentelemetry-exporter-otlp = "^1.31.1"
opentelemetry-instrumentation = "^0.52b1"
opentelemetry-exporter-otlp-proto-http = "^1.31.1"
apscheduler = "^3.11.0"
starlette = "^0.46.1"
dotenv = "^0.9.9"
uvicorn = "^0.34.0"
openai = "1.97.1"
r2r = "^3.5.16"
atlassian-python-api = "^4.0.4"
cryptography = "^46.0.1"
email-validator = "^2.3.0"
python-multipart = "^0.0.20"
fastapi-sso = "^0.18.0"

[tool.poetry.group.checks.dependencies]
bandit = "^1.7.9"
coverage = "^7.5.4"
mypy = "^1.10.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"
pandera = { extras = ["mypy"], version = "^0.20.1" }
ruff = "^0.5.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^1.0.0"


[tool.poetry.group.commits.dependencies]
commitizen = "^3.27.0"
pre-commit = "^3.7.1"

[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.5.1"

[tool.poetry.group.notebooks.dependencies]
ipykernel = "^6.29.4"
nbformat = "^5.10.4"

# CONFIGURATIONS

[tool.bandit]
targets = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "poetry"
changelog_start_rev = "v1.0.0"
update_changelog_on_bump = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["__main__.py"]

[tool.mypy]
pretty = true
strict = false
python_version = "3.11"
check_untyped_defs = false
ignore_missing_imports = true
disable_error_code = ["return-value"]
plugins = ["pandera.mypy", "pydantic.mypy"]
exclude = ["src/fastapi_autogen_team/__init__.py"]

[tool.pytest.ini_options]
addopts = "--verbosity=2"
pythonpath = ["src"]

[tool.ruff]
fix = true
indent-width = 4
line-length = 120
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]

# SYSTEMS

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

